DismissEvent

+++ src/main/java/jp/co/youmeee/clipper/RectClipEntry.kt
@@ -10,8 +10,8 @@ class RectClipEntry(override val targetView: View) : ClipEntry() {
     private val targetGlobalVisibleRect = Rect()
     private var clipMargin: Int = 0
 
-    constructor(target: View, clipMargin: Float) : this(target) {
-        this.clipMargin = clipMargin.toInt()
+    constructor(target: View, clipMargin: Int) : this(target) {
+        this.clipMargin = clipMargin
     }
 
     override fun clip(canvas: Canvas, paint: Paint, decorRect: Rect) {
